cleanup

+++ api/models.py
@@ -529,7 +529,7 @@ class User(Model):
                     notes=data['status']['notes'] if 'notes' in data['status'] else None,
                 )
             ids.append(user_status.id)
-        UserStatus.objects.get_queryset().exclude(id__in=ids).delete()
+        UserStatus.objects.get_queryset().filter(user_id=self.id).exclude(id__in=ids).delete()
         return self
 
     def update_status_attachments(self, data):
@@ -556,7 +556,7 @@ class User(Model):
                             position=attachment['position'] if 'position' in attachment else 0,
                         )
                     ids.append(user_status_attachment.id)
-        UserStatusAttachment.objects.get_queryset().exclude(id__in=ids).delete()
+        UserStatusAttachment.objects.get_queryset().filter(user_status__user_id=self.id)exclude(id__in=ids).delete()
         return self
 
     def update_urls(self, data):
